NativeMemory objects to use `nint/nuint`

+++ Automata.Engine/Rendering/OpenGL/Memory/NativeMemoryManager.cs
@@ -6,23 +6,23 @@ namespace Automata.Engine.Rendering.OpenGL.Memory
     internal sealed unsafe class NativeMemoryManager<T> : MemoryManager<T> where T : unmanaged
     {
         private readonly T* _Pointer;
-        private readonly uint _Length;
+        private readonly int _Length;
         private readonly Memory<T> _Memory;
 
-        public NativeMemoryManager(T* pointer, uint length)
+        public NativeMemoryManager(T* pointer, int length)
         {
             _Pointer = pointer;
             _Length = length;
-            _Memory = CreateMemory((int)length);
+            _Memory = CreateMemory(length);
         }
 
-        public Memory<T> Slice(uint start) => _Memory.Slice((int)start);
-        public Memory<T> Slice(uint start, uint length) => _Memory.Slice((int)start, (int)length);
+        public Memory<T> Slice(int start) => _Memory.Slice(start);
+        public Memory<T> Slice(int start, int length) => _Memory.Slice(start, length);
 
 
         #region MemoryManager
 
-        public override Span<T> GetSpan() => new Span<T>(_Pointer, (int)_Length);
+        public override Span<T> GetSpan() => new Span<T>(_Pointer, _Length);
 
         public override MemoryHandle Pin(int elementIndex = 0)
         {
